requirements.txt
pandas
numpy
apache-airflow
beautifulsoup4
opencv-python